Seperated out the common methods and properties of the query builder libraries into a generic QB lib that the driver-specific QBs extend.
