+++ drivers/mssql/query_builder.js
@@ -1,174 +1,14 @@
-class QueryBuilder {
+const GenericQueryBuilder = require('../query_builder.js');
+
+class QueryBuilder extends GenericQueryBuilder {
     constructor() {
-        this.where_array = [];
-        this.where_in_array = [];
-        this.from_array = [];
-        this.join_array = [];
-        this.select_array = [];
-        this.set_array = [];            // has to be array to work as reference
-        this.order_by_array = [];
-        this.group_by_array = [];
-        this.having_array = [];
-        this.limit_to = [];             // has to be array to work as reference
-        this.offset_val = [];           // has to be array to work as reference
-        this.join_clause = [];
-        this.last_query_string = [];    // has to be array to work as reference
-        this.distinct_clause = [];      // has to be array to work as reference
-        this.aliased_tables = [];
-    }
-
-     // ------------------------------ GENERIC METHODS ------------------------------//
-    _array_values(item) {
-        const keys = Object.keys(item);
-        const length = keys.length;
-        const values = Array(length);
-        for (let i = 0; i < length; i++) {
-          values[i] = item[keys[i]];
-        }
-        return values;
-    }
-
-    _prepare_for_limit_and_offset(item, type = 'limit') {
-        type = type.toLowerCase();
-
-        if (!/^(string|number)$/.test(typeof item)) {
-            throw new Error("Only integers or integers in the form of a string are allowed");
-        }
-
-        if (typeof item === 'string') {
-            item = item.trim();
-            if (!/^\d+$/.test(item)) {
-                throw new Error("The string you provided to " + type + " by contains non-integer values--this isn't allowed.");
-            }
-            // Force to an integer
-            item = parseInt(item);
-        }
-
-        // Make sure the number is a good one
-        if (typeof item === 'number') {
-            // Check for NaN and Infinity
-            if (item !== +item || item === Infinity) {
-                throw new Error("You have not provided a valid number to " + type + " by!");
-            }
-
-            // Make sure it's positive
-            if (item < 0) {
-                throw new Error("Only positive integeres are allowed when " + (type == 'offset' ? 'OFFSET' : 'LIMIT') + "ing SQL results!");
-            }
-
-            // Only allow integers
-            if (item % 1 !== 0) {
-                throw new Error("You cannot " + type + " a SQL result set with a floating point value!");
-            }
-        }
-        else {
-            throw new Error("There was an unrecoverable error while parsing the value provdided in your " + type + " statement.");
-        }
-
-        return item;
-    }
-
-    _extract_having_parts(key, key_array) {
-        let m;
-        key = key.trim().replace(/\s+/g,' ');
-        const str_condition  = /^([^\s]+\s(<=|>=|<>|>|<|!=|=))+\s"([^"]+)"$/;  //"// had to do this for syntax highlighting
-        const sstr_condition = /^([^\s]+\s(<=|>=|<>|>|<|!=|=))+\s'([^']+)'$/;  //'// had to do this for syntax highlighting
-        const num_condition  = /^([^\s]+\s(<=|>=|<>|>|<|!=|=))+\s((?=.)([+-]?([0-9]*)(\.([0-9]+))?))$/;
-        const bool_condition = /^([^\s]+\s(<=|>=|<>|>|<|!=|=))+\s((true|false)+)$/;
-
-        if (m = str_condition.exec(key)) {
-            key_array[m[1]] = m[3];
-            key = key_array;
-        }
-        else if (m = sstr_condition.exec(key)) {
-            //console.log("Key has sstring value");
-            key_array[m[1]] = m[3];
-            key = key_array;
-        }
-        else if (m = num_condition.exec(key)) {
-            //console.log("Key has numeric value");
-            key_array[m[1]] = m[3];
-            key = key_array;
-        }
-        else if (m = bool_condition.exec(key)) {
-            //console.log("Key has boolean value");
-            key_array[m[1]] = m[3];
-            key = key_array;
-        }
-        else {
-            throw new Error("An invalid condition was supplied (" + key + ") in your having statement!");
-        }
-
-        return key_array;
-    }
-
-    // Simply setting all properties to [] causes reference issues in the parent class.
-    _clear_array(a, debug) {
-        if (debug === true) {
-            console.log("DEBUG before (" + Object.prototype.toString.call(a) + "):");
-            console.dir(a);
-        }
-        if (Object.prototype.toString.call(a) === Object.prototype.toString.call({})) {
-            for (let key in a) {
-                if (a.hasOwnProperty(key)) {
-                    delete a[key];
-                }
-            }
-        }
-        else if (Array.isArray(a)) {
-            while (a.length > 0) {
-                a.pop();
-            }
-        }
-        if (debug === true) {
-            console.log("DEBUG after (" + Object.prototype.toString.call(a) + "):");
-            console.dir(a);
-        }
+        super();
+        this.rand_word = 'NEWID()';
     }
 
     // ---------------------------------------- SQL ESCAPE FUNCTIONS ------------------------ //
-    _track_aliases(table) {
-        if (Object.prototype.toString.call(table) === Object.prototype.toString.call({})) {
-            for (let i in table) {
-                const t = table[i];
-                this._track_aliases(t);
-            }
-            return;
-        }
-
-        // Does the string contain a comma?  If so, we need to separate
-        // the string into discreet statements
-        if (table.indexOf(',') !== -1) {
-            return this._track_aliases(table.split(','));
-        }
-
-        // if a table alias is used we can recognize it by a space
-        if (table.indexOf(' ') !== -1) {
-            // if the alias is written with the AS keyword, remove it
-            table = table.replace(/\s+AS\s+/gi, ' ');
-
-            // Grab the alias
-            const alias = table.slice(table.lastIndexOf(' ')).trim().replace(/`/g,'');
-
-            // Store the alias, if it doesn't already exist
-            if (this.aliased_tables.indexOf(alias) == -1) {
-                this.aliased_tables.push(alias);
-            }
-        }
-    }
-
-    _create_aliases_from_table(item) {
-        if (item.indexOf('.') !== -1) {
-            return item.split('.').reverse()[0];
-        }
-
-        return item;
-    }
-
     _escape_identifiers(item = '*') {
-        if (item === '*') {
-            return item;
-        }
+        if (item === '*') return item;
 
         if (Object.prototype.toString.call(item) === Object.prototype.toString.call({})) {
             for (let i in item) {
@@ -176,12 +16,12 @@ class QueryBuilder {
             }
             return item;
         }
-        else if ((typeof item === 'string' && /^\d+$/.test(item)) || item[0] === "'" || item[0] === '"' || item.indexOf('(') !== -1) {
+        else if ((typeof item === 'string' && /^\d+$/.test(item)) || item[0] === "'" || item.indexOf('(') !== -1) {
             return item;
         }
 
         let str;
-        if (item.indexOf('.' + '*') !== -1) {
+        if (item.indexOf('.*') !== -1) {
             str = item.replace(/`?([^`\.]+)`?\./ig, "`$1`.");
         }
         else {
@@ -285,16 +125,6 @@ class QueryBuilder {
         return item + alias;
     };
 
-    _has_operator(str) {
-        if (typeof str === 'string' && str.length > 0) {
-            const match = /(<|>|!|=|\sIS NULL|\sIS NOT NULL|\sEXISTS|\sBETWEEN|\sLIKE|\sCASE|\sTHEN|\sWHEN|\sIN\s*\(|\s)/i.test(str.trim());
-            if (!match) {
-                return false;
-            }
-        }
-        return true;
-    };
-
     _qb_escape(str) {
         const mysql = require('mysql');
 
@@ -309,60 +139,7 @@ class QueryBuilder {
         return str;
     };
 
-
     // ---------------------------- SQL BUILD TOOLS ----------------------------//
-    _build_where_clause() {
-        let sql = '';
-        if (this.where_array.length > 0) {
-            sql += " WHERE ";
-        }
-        sql += this.where_array.join(" ");
-        return sql;
-    };
-
-    _build_from_clause() {
-        let sql = '';
-        if (this.from_array.length > 0) {
-            sql += " FROM ";
-        } else {
-            throw new Error("You have not provided any tables, views, or store procedures for this query!!");
-        }
-        sql += this.from_array.join(', ');
-        return sql;
-    };
-
-    _build_join_string() {
-        let sql = '';
-        sql += this.join_array.join(' ');
-        if (sql.length > 0) sql = ' ' + sql;
-        return sql;
-    };
-
-    _build_group_by_clause() {
-        if (this.group_by_array.length <= 0) return '';
-
-        let sql = ' GROUP BY ';
-        sql += this.group_by_array.join(', ');
-        return sql;
-    };
-
-    _build_having_clause() {
-        if (this.having_array.length <= 0) return '';
-
-        let sql = ' HAVING ';
-        sql += this.having_array.join(' ');
-        return sql;
-    };
-
-    _build_order_by_clause() {
-        if (this.order_by_array.length <= 0) return '';
-
-        let sql = ' ORDER BY ';
-        sql += this.order_by_array.join(', ');
-
-        return sql;
-    };
-
     _build_limit_clause(sql, limit, offset) {
         if (!limit) return sql;
 
@@ -377,29 +154,6 @@ class QueryBuilder {
         return sql.replace(/\s+$/, ' ') + 'LIMIT ' + offset + limit;
     };
 
-    _compile_select() {
-        const distinct_clause = this.distinct_clause[0] || '';
-        let sql = 'SELECT ' + distinct_clause;
-        if (this.select_array.length === 0) {
-            sql += '*';
-        } else {
-            sql += this.select_array.join(', ');
-        }
-
-        sql += this._build_from_clause()
-            + this._build_join_string()
-            + this._build_where_clause()
-            + this._build_group_by_clause()
-            + this._build_having_clause()
-            + this._build_order_by_clause();
-
-        const limit_to = this.limit_to[0] || false;
-        const offset_val = this.offset_val[0] || false;
-
-        sql = this._build_limit_clause(sql, limit_to, offset_val);
-        return sql;
-    };
-
     _compile_delete() {
         if (this.from_array.length === 0) {
             throw new Error('You have not specified any tables to delete from!');
@@ -412,34 +166,6 @@ class QueryBuilder {
         const offset_val = this.offset_val[0] || false;
 
         const sql = 'DELETE' + this._build_from_clause() + this._build_where_clause();
-        return this._build_limit_clause(sql, limit_to, offset_val);
-    };
-
-    _compile_update() {
-        const valstr = [];
-        for (let i in this.set_array) {
-            const key = Object.keys(this.set_array[i])[0];
-            const val = this.set_array[i][key];
-            valstr.push(key + ' = ' + val);
-        }
-
-        if (this.from_array.length !== 1) {
-            if (this.from_array.length === 0) {
-                throw new Error("You haven't provided any tables to build UPDATE query with!");
-                return '';
-            }
-            throw new Error("You have provided too many tables to build UPDATE query with!");
-            return '';
-        }
-
-        const table = this.from_array.toString();
-
-        const limit_to = this.limit_to[0] || false;
-        const offset_val = this.offset_val[0] || false;
-
-        let sql = `UPDATE (${table}) SET ${valstr.join(', ')}`;
-        sql += this._build_where_clause();
-        sql += this._build_order_by_clause();
         return this._build_limit_clause(sql, limit_to, offset_val);
     };
 
@@ -471,298 +197,56 @@ class QueryBuilder {
         return verb + `INTO ${this.from_array[0]} (${keys.join(', ')}) VALUES (${values.join(', ')})${suffix}`;
     };
 
-    reset_query(new_last_query,debug) {
-        this._clear_array(this.where_array, debug);
-        this._clear_array(this.where_in_array);
-        this._clear_array(this.from_array);
-        this._clear_array(this.join_array);
-        this._clear_array(this.select_array);
-        this._clear_array(this.set_array);
-        this._clear_array(this.order_by_array);
-        this._clear_array(this.group_by_array);
-        this._clear_array(this.having_array);
-        this._clear_array(this.limit_to);
-        this._clear_array(this.offset_val);
-        this._clear_array(this.join_clause);
-        this._clear_array(this.distinct_clause);
-        this._clear_array(this.aliased_tables);
-
-        this._clear_array(this.last_query_string);
-        if (typeof new_last_query === 'string') {
-            this.last_query_string.push(new_last_query);
-        }
-    }
-
-    where(key, value = null, escape) {
-        if (Object.prototype.toString.call(key) === Object.prototype.toString.call({}) && typeof value === 'boolean') {
-            escape = (typeof escape === 'boolean' ? escape : value);
+    _compile_select() {
+        const distinct_clause = this.distinct_clause[0] || '';
+        let sql = 'SELECT ' + distinct_clause;
+        if (this.select_array.length === 0) {
+            sql += '*';
+        } else {
+            sql += this.select_array.join(', ');
         }
 
-        escape = (typeof escape === 'boolean' ? escape : true);
+        sql += this._build_from_clause()
+            + this._build_join_string()
+            + this._build_where_clause()
+            + this._build_group_by_clause()
+            + this._build_having_clause()
+            + this._build_order_by_clause();
 
-        if (typeof key === 'string' && typeof value === 'object' && Array.isArray(value) && value.length > 0) {
-            return this._where_in(key, value, false, 'AND ');
-        }
-        return this._where(key, value, 'AND ', escape);
-    }
+        const limit_to = this.limit_to[0] || false;
+        const offset_val = this.offset_val[0] || false;
 
-    or_where(key, value=null, escape) {
-        escape = (typeof escape === 'boolean' ? escape : true);
+        sql = this._build_limit_clause(sql, limit_to, offset_val);
+        return sql;
+    };
 
-        if (typeof key === 'string' && typeof value === 'object' && Array.isArray(value) && value.length > 0) {
-            return this._where_in(key, value, false, 'OR ');
-        }
-        return this._where(key, value, 'OR ', escape);
-    }
-
-    _where(key, value=null, type='AND ', escape) {
-        escape = (typeof escape === 'boolean' ? escape : true);
-
-        // Must be an object or a string
-        if (Object.prototype.toString.call(key) !== Object.prototype.toString.call({})) {
-            // If it's not an object, it must be a string
-            if (typeof key !== 'string') {
-                throw new Error("where(): If first parameter is not an object, it must be a string. " + typeof key + " provided.");
-            } else {
-                // If it is a string, it can't be an empty one
-                if (key.length == 0) {
-                    throw new Error("where(): No field name or query provided!");
-                }
-            }
-
-            // If it's a actual where clause string (with no paranthesis),
-            // not just a field name, split it into individual parts to escape it properly
-            if (/(<=|>=|<>|>|<|!=|=)/.test(key) && key.indexOf('(') === -1 && escape === true) {
-                const filters = key.split(/\s+(AND|OR)\s+/i);
-                if (filters.length > 1) {
-                    const self = this;
-                    const parse_statement = (statement, joiner) => {
-                        const parsed = statement.match(/^([^<>=!]+)(<=|>=|<>|>|<|!=|=)(.*)$/);
-                        if (parsed.length >= 4) {
-                            const key = parsed[1].trim() + (parsed[2].trim() !== '=' ? ' ' + parsed[2].trim() : '');
-                            const value = parsed[3].trim().replace(/^((?:'|"){1})(.*)/, "$2").replace(/'$/,'');
-                            if (joiner === null || /AND/i.test(joiner)) {
-                                self.where(key, value, true);
-                            } else {
-                                self.or_where(key, value, true);
-                            }
-                        }
-                    };
-                    parse_statement(filters.shift(), null);
-                    while (filters.length > 0) {
-                        const joiner = filters.shift();
-                        const statement = filters.shift();
-                        parse_statement(statement, joiner);
-                    }
-                    return this;
-                }
-            }
-
-            const key_array = {};
-            key_array[key] = value;
-            key = key_array;
+    _compile_update() {
+        const valstr = [];
+        for (let i in this.set_array) {
+            const key = Object.keys(this.set_array[i])[0];
+            const val = this.set_array[i][key];
+            valstr.push(key + ' = ' + val);
         }
 
-        if (Object.keys(key).length == 0) {
-            throw new Error("where(): You haven't provided any key value pairs to limit the resultset by.");
+        if (this.from_array.length !== 1) {
+            if (this.from_array.length === 0) {
+                throw new Error("You haven't provided any tables to build UPDATE query with!");
+                return '';
+            }
+            throw new Error("You have provided too many tables to build UPDATE query with!");
+            return '';
         }
 
-        for (let k in key) {
-            let v = key[k];
+        const table = this.from_array.toString();
 
-            if (typeof v === 'object' && Array.isArray(v) && v.length > 0) {
-                return this._where_in(k, v, false, type, escape);
-            }
+        const limit_to = this.limit_to[0] || false;
+        const offset_val = this.offset_val[0] || false;
 
-            const prefix = (this.where_array.length == 0 ? '' : type);
-
-            if (v === null && !this._has_operator(k)) {
-                k += ' IS NULL';
-            }
-
-            if (v !== null) {
-                if (escape === true) {
-                    k = this._protect_identifiers(k,escape);
-                    v = ' ' + this._qb_escape(v);
-                }
-
-                if (escape === false && Object.prototype.toString.call(key) === Object.prototype.toString.call({})) {
-                    v = ' ' + this._qb_escape(v);
-                }
-
-                if (!this._has_operator(k)) {
-                    k += ' =';
-                }
-            }
-            else {
-                k = this._protect_identifiers(k,escape);
-            }
-
-            if (v) {
-                this.where_array.push(prefix+k+v);
-            }
-            else {
-                this.where_array.push(prefix+k);
-            }
-        }
-
-        return this;
-    }
-
-    where_in(key, values, escape) {
-        return this._where_in(key, values, false, 'AND ', escape);
-    }
-
-    or_where_in(key, values, escape) {
-        return this._where_in(key, values, false, 'OR ', escape);
-    }
-
-    where_not_in(key, values, escape) {
-        return this._where_in(key, values, true, 'AND ', escape);
-    }
-
-    or_where_not_in(key, values, escape) {
-        return this._where_in(key, values, true, 'OR ', escape);
-    }
-
-    _where_in(key='', values=[], not, type='AND ', escape) {
-        not = (not ? ' NOT' : '');
-        escape = (typeof escape === 'boolean' ? escape : true);
-
-        if (typeof key !== 'string' || (typeof key === 'string' && key.length == 0)) {
-            throw new Error("where_" + (not === '' ? '' : not.toLowerCase() + '_') + "in(): Invalid field name provided.");
-        }
-
-        // Values must be an array...
-        if (!Array.isArray(values)) {
-            throw new Error("where_" + (not === '' ? '' : not.toLowerCase() + '_') + "in(): Invalid second parameter provided--it must be an array of scalar values.");
-        }
-        else {
-            if (values.length == 0) {
-                throw new Error("where_" + (not === '' ? '' : not.toLowerCase() + '_') + "in(): You have provided an empty list of values to limit resultset by.");
-            }
-        }
-
-        for (let i in values) {
-            this.where_in_array.push(this._qb_escape(values[i]));
-        }
-
-        const prefix = (this.where_array.length == 0 ? '' : type);
-        const where_in = prefix + this._protect_identifiers(key,escape) + not + " IN (" + this.where_in_array.join(', ') + ")";
-        this.where_array.push(where_in);
-
-        // reset the array for multiple calls
-        this._clear_array(this.where_in_array);
-        return this;
-    }
-
-    like(field, match, side) {
-        return this._like(field, match, 'AND ', side, '');
-    }
-
-    not_like(field, match, side) {
-        return this._like(field, match, 'AND ', side, ' NOT');
-    }
-
-    or_like(field, match, side) {
-        return this._like(field, match, 'OR ', side, '');
-    }
-
-    or_not_like(field, match, side) {
-        return this._like(field, match, 'OR ', side, ' NOT');
-    }
-
-    _like(field, match, type='AND ', side='both', not='') {
-        match = (/^(string|number|boolean)$/.test(typeof match) ? match : null);
-
-        if (typeof field === 'string' && field.length == 0) {
-            throw new Error("like(): The field you provided is empty.");
-        }
-        else if (typeof field === 'object' && (field.length == 0 || Object.keys(field).length === 0)) {
-            throw new Error("like(): The object you provided is empty.");
-        }
-        else if (!/^(string|object)$/.test(typeof field)) {
-            throw new Error("like(): You have provided an invalid value as the first parameter. Only valid strings and objects are allowed.");
-        }
-
-        if (Object.prototype.toString.call(field) !== Object.prototype.toString.call({})) {
-            if (match === null) {
-                throw new Error("like(): Since your first parameter is a string, your second param must a valid number, boolean, or string.");
-            }
-
-            const field_array = {};
-            field_array[field] = match;
-            field = field_array;
-        }
-
-        for (let k in field) {
-            let like_statement;
-            const v = field[k];
-            k = this._protect_identifiers(k.trim());
-
-            // Make sure value is only string, number, or boolean
-            if (!/^(string|number|boolean)$/.test(typeof v)) {
-                throw new Error("like(): You have provided an invalid value as the second parameter. Only valid strings, numbers, and booleans are allowed.");
-            }
-            // If number, don't allow Infinity or NaN
-            else if (typeof v === 'number' && (v === Infinity || (v !== +v))) {
-                throw new Error("like(): You have provided an invalid number value as the second parameter. Only valid strings, numbers, and booleans are allowed.");
-            }
-
-            if (side === 'none') {
-                like_statement =  k + not + ' LIKE ' + "'" + v + "'";
-            }
-            else if (side === 'before' || side === 'left') {
-                like_statement = k + not + ' LIKE ' + "'%" + v + "'";
-            }
-            else if (side === 'after' || side === 'right') {
-                like_statement = k + not + ' LIKE ' + "'" + v + "%'";
-            }
-            else if (side === 'both') {
-                like_statement = k + not + ' LIKE ' + "'%" + v + "%'";
-            }
-            else {
-                throw new Error("like(): Invalid direction provided!");
-            }
-
-            this._where(like_statement, null, type, false);
-        }
-
-        return this;
-    }
-
-    from(from_param) {
-        if (!Array.isArray(from_param)) {
-            from_param = [from_param];
-        }
-        for (let i in from_param) {
-            let val = from_param[i];
-
-            if (typeof val !== 'string' || val.trim() === '') continue;
-
-            if (val.indexOf(',') !== -1) {
-                const objects = val.split(',');
-                for (let j in objects) {
-                    const v = objects[j].trim();
-
-                    this._track_aliases(v);
-
-                    this.from_array.push(this._protect_identifiers(v, true));
-                }
-            }
-            else {
-                val = val.trim();
-
-                // Extract any aliases that might exist.  We use this information
-                // in the protect_identifiers function to know whether to add a table prefix
-                this._track_aliases(val);
-
-                this.from_array.push(this._protect_identifiers(val, true));
-            }
-        }
-
-        return this;
-    }
+        let sql = `UPDATE (${table}) SET ${valstr.join(', ')}`;
+        sql += this._build_where_clause();
+        sql += this._build_order_by_clause();
+        return this._build_limit_clause(sql, limit_to, offset_val);
+    };
 
     join(table, relation, direction, escape) {
         if (typeof table !== 'string' || table.trim().length == 0) {
@@ -838,431 +322,6 @@ class QueryBuilder {
         return this;
     }
 
-    select(select,escape) {
-        // First param must be a non-empty string or array
-        if (typeof select === 'string') {
-            select = select.trim();
-            if (select.length == 0) {
-                throw new Error("Your select string is empty!");
-            }
-        }
-        else if (Array.isArray(select)) {
-            if (select.length == 0) {
-                throw new Error("Your select array is empty!");
-            }
-        }
-        else {
-            throw new Error("Select method requires a string or array to be passed in the first parameter!");
-        }
-
-        if (typeof escape !== 'boolean') escape = true;
-
-        // Split statements out into individual ones by comma (unless there is a function or subquery with commas in it)
-        if (typeof select === 'string') {
-            if (select.indexOf(')') === -1) {
-                select = select.split(',');
-            } else {
-                if (escape === true) {
-                    // Prevent it from trying to parse select statements with functions and if statements
-                    if (/\w\s?\(/.test(select)) throw new Error("Select statements with subqueries or functions cannot be escaped! Please escape manually and pass FALSE as the second paramter to the select method.");
-
-                    // Identify individual statements within select string
-                    let m, open_paren_index, inner_parenthesis;
-                    const reg = /\)/g;
-                    while ((m = reg.exec(select) !== null)) {
-                        open_paren_index = m.input.substring(0, m.index).lastIndexOf('(');
-                        if (open_paren_index !== -1) {
-                            inner_parenthesis = m.input.substring((open_paren_index + 1), m.index);
-                            if (inner.parenthesis.indexOf(',') !== -1) {
-                                throw new Error("Select statements with subqueries or functions cannot be escaped! Please escape manually and pass FALSE as the second paramter to the select method.");
-                                break;
-                            }
-                        }
-                    }
-                }
-                else {
-                    select = [select];
-                }
-            }
-        }
-
-        for (let i in select) {
-            const val = select[i].trim();
-
-            if (val !== '') {
-                this.select_array.push(this._protect_identifiers(val, escape));
-            }
-        }
-        return this;
-    }
-
-    select_min(select,alias) {
-        return this._min_max_avg_sum(select, alias, 'MIN');
-    }
-
-    select_max(select,alias) {
-        return this._min_max_avg_sum(select, alias, 'MAX');
-    }
-
-    select_avg(select,alias) {
-        return this._min_max_avg_sum(select, alias, 'AVG');
-    }
-
-    select_sum(select,alias) {
-        return this._min_max_avg_sum(select, alias, 'SUM');
-    }
-
-    _min_max_avg_sum(select='', alias='', type='MAX') {
-
-        if (typeof select !== 'string' || select === '') {
-            throw Error("Invalid query!");
-            return this;
-        }
-
-        type = type.toUpperCase();
-
-        if (['MAX', 'MIN', 'AVG', 'SUM'].indexOf(type) === -1) {
-            throw Error("Invalid function type!");
-            return this;
-        }
-
-        if (alias == '') {
-            alias = this._create_aliases_from_table(select.trim());
-        }
-
-        const sql = `${type}(${this._protect_identifiers(select.trim())}) AS ${alias}`;
-
-        this.select_array.push(sql);
-
-        return this;
-    }
-
-    distinct(do_distinct) {
-        do_distinct = (typeof do_distinct !== 'boolean' ? true : do_distinct);
-
-        if (do_distinct) {
-            this._clear_array(this.distinct_clause);
-            this.distinct_clause.push('DISTINCT ');
-        }
-        else {
-            this._clear_array(this.distinct_clause);
-        }
-        return this;
-    }
-
-    group_by(by) {
-        if (typeof by === 'string') {
-            by = by.trim();
-            if (by.length <= 0) {
-                throw new Error("You haven't provided any fields to group by!");
-            }
-            by = by.split(',');
-        }
-
-        if (!Array.isArray(by)) {
-            throw new Error("You have provided an invalid value to the group_by() method. Only strings and arrays of strings are allowed.");
-        }
-
-        if (by.length <= 0) {
-            throw new Error("You haven't provided any fields to group by!");
-        }
-
-        for (let key in by) {
-            if (typeof by[key] !== 'string') {
-                throw new Error("You have provided an invalid value to the group_by() method. Only strings and arrays of strings are allowed!");
-            }
-
-            const val = by[key].trim();
-
-            if (val !== '') {
-                this.group_by_array.push(this._protect_identifiers(val));
-            }
-        }
-        return this;
-    }
-
-    having(key, value, escape) {
-        escape = (typeof escape !== 'boolean' ? true : escape);
-        return this._having(key, value, 'AND ', escape);
-    }
-
-    or_having(key, value, escape) {
-        escape = (typeof escape !== 'boolean' ? true : escape);
-        return this._having(key, value, 'OR ', escape);
-    }
-
-    _having(key, value, type='AND ', escape) {
-
-        let m;
-        let key_array = {};
-        const key_is_object = Object.prototype.toString.call(key) === Object.prototype.toString.call({});
-        const key_is_array = Array.isArray(key);
-
-        if (/^(string|number|boolean)$/.test(typeof value)) { // if the value is a string, number, or boolean...
-            if (typeof key !== 'string' || /^\W+$/i.test(key)) { // if the key is not a string...
-                throw new Error("having(): The value you provided when calling having() will be ignored since the first parameter is not a single field provided in string form.");
-            }
-            key_array[key] = value;
-            key = key_array;
-        }
-        else if (typeof value === 'undefined' || value === null) {
-            if (key_is_object === false) {
-                if (typeof key === 'string') {
-                    if (value === null) {
-                        key_array[key] = null;
-                        key = key_array;
-                    }
-                    else {
-                        key = this._extract_having_parts(key, key_array);
-                    }
-                }
-                else if (key_is_array === true) {
-                    //console.log("Key is NOT a string");
-                    for (let i in key) {
-                        if (typeof key[i] !== 'string') {
-                            throw new Error("having(): You've provided an unparseable format to the having() method..");
-                        }
-                        else {
-                            key_array = this._extract_having_parts(key[i], key_array);
-                        }
-                    }
-                    key = key_array;
-                }
-            }
-        }
-        else {
-            throw new Error(`Improper value supplied as the second parameter (${typeof value}) of the having() method.`);
-        }
-
-
-        for (let k in key) {
-            let v = key[k];
-            const prefix = (this.having_array.length == 0 ? '' : type);
-
-            if (escape === true) {
-                k = this._protect_identifiers(k);
-            }
-
-            if (v === null) {
-                k += ' IS';
-            }
-            else if (!this._has_operator(k)) {
-                k += ' =';
-            }
-
-            if (v != '') {
-                v = ' ' + this._qb_escape(v);
-            }
-
-            this.having_array.push(prefix + k + v);
-        }
-
-        return this;
-    }
-
-    order_by(orderby, direction) {
-        let m;
-        const rand_word = 'RAND()';
-        direction = (typeof direction === 'string' ? direction.toLowerCase().trim() : '');
-
-        // Don't need to do anything below if the direction provided is random
-        if ((direction === 'random' || direction === 'rand' || direction === 'rand()')) {
-            this.order_by_array.push(rand_word);
-            return this;
-        }
-
-        // normalize orderby to be an array of items
-        if (!Array.isArray(orderby)) {
-            if (typeof orderby === 'string') {
-                orderby = orderby.trim();
-                if (orderby.length == 0) {
-                    throw new Error("You haven't provided any fields to order by!!");
-                }
-                orderby = orderby.split(',');
-            } else if (!orderby && (/(random|RAND|RAND\(\))/i.test(direction))) {
-                this.order_by_array.push(rand_word);
-                return this;
-            }
-            else {
-                throw new Error("No field provided to order by!");
-            }
-        }
-
-        if (orderby.length === 0) {
-            throw new Error("You haven't provided any fields to order by!");
-        }
-
-        for (let i in orderby) {
-            orderby[i] = orderby[i].replace(/\s+/g, ' ');
-
-            if (m = orderby[i].match(/([^\s]+)\s+(ASC|DESC|RAND\(\))/i)) {
-                if (m[2].trim() === 'RAND()') {
-                    this.order_by_array.push(rand_word);
-                    return this;
-                }
-                orderby[i] = {field: this._protect_identifiers(m[1].trim()), direction: m[2].trim().toUpperCase()};
-            } else {
-                if (/^(ASC|DESC)$/i.test(direction) || direction === '') {
-                    orderby[i] = {field: this._protect_identifiers(orderby[i].trim()), direction: (direction !== '' ? direction.toUpperCase() : 'ASC')};
-                } else {
-                    throw new Error("Invalid direction provided in order_by method! Only 'ASC', 'DESC', and 'RAND' are allowed!");
-                }
-            }
-
-            this.order_by_array.push(orderby[i].field + ' ' + orderby[i].direction);
-        }
-
-        return this;
-    }
-
-    limit(limit, offset) {
-        this._clear_array(this.limit_to);
-        this.limit_to.push(this._prepare_for_limit_and_offset(limit, 'limit'));
-
-        if (offset !== undefined) {
-            return this.offset(offset);
-        }
-
-        return this;
-    }
-
-    offset(offset) {
-        this._clear_array(this.offset_val);
-        this.offset_val.push(this._prepare_for_limit_and_offset(offset, 'offset'));
-        return this;
-    }
-
-    set(key, value, escape) {
-        escape = (typeof escape === 'boolean' ? escape : true);
-
-        if (typeof key === 'string') {
-            // Convert key and value params to {key: value}
-            key = key.trim();
-            if (key.length == 0) throw new Error("set(): Invalid field name provided!");
-
-            if (typeof value === 'undefined')
-                throw new Error("set(): First param was string but no value (second param) provided to set!");
-
-            const key_array = {};
-            key_array[key] = value;
-            key = key_array;
-        }
-        else if (Object.prototype.toString.call(key) === Object.prototype.toString.call({})) {
-            if (Object.keys(key).length === 0) {
-                throw new Error("set(): The object you provided is empty.");
-            }
-
-            if (typeof value !== 'undefined' & value !== null) {
-                throw new Error("set(): The value you provided in the second parameter will be ignored since you passed an object as the first parameter.");
-            }
-        }
-        else {
-            throw new Error("set(): First parameter must be a non-empty string or non-empty object! " + typeof key + " provided.");
-        }
-
-
-        // Add each key:value pair to the set_array
-        for (let i in key) {
-            let v = key[i];
-            if (typeof v === 'undefined') continue;
-
-            if (v instanceof Date) v = v.toString();
-
-            if (!/^(number|string|boolean)$/.test(typeof v) && v !== null) {
-                throw new Error("set(): Invalid value provided! (provided: " + v + " (type: " + (typeof v) + ")");
-            }
-            else if (typeof v === 'number' && (v === Infinity || v !== +v)) {
-                throw new Error("set(): Infinity and NaN are not valid values in MySQL!");
-            }
-
-            // Escape the key to be DRY
-            const escaped_key = this._protect_identifiers(i,escape);
-
-            // Build a temporary object with escaped key and val
-            const temp = {};
-            if (escape === false) {
-                temp[escaped_key] = v;
-            } else {
-                temp[escaped_key] = this._qb_escape(v);
-            }
-
-            // Determine if this key has already been set
-            let found_index = null;
-            for (let j in this.set_array) {
-                if (this.set_array[j].hasOwnProperty(escaped_key)) {
-                    found_index = j;
-                    break;
-                }
-            }
-
-            // Update value if key already set or add if not found
-            if (found_index !== null) {
-                this.set_array[found_index] = temp;
-            } else {
-                this.set_array.push(temp);
-            }
-        }
-
-        //console.dir(this.set_array);
-
-        return this;
-    }
-
-    insert(table, set, ignore, suffix) {
-        return this._insert(table, set, ignore, suffix);
-    }
-
-    _insert(table, set, ignore, suffix) {
-        table = table || ''
-        ignore = (typeof ignore !== 'boolean' ? false : ignore);
-        suffix = (typeof suffix !== 'string' ? '' : ' ' + suffix);
-
-        if (/^(number|boolean)$/.test(typeof set) || (typeof set == 'string' && set !== '') || Object.prototype.toString.call(set) === Object.prototype.toString.call(/test/)) {
-            throw new Error("insert(): Invalid data provided to insert into database!");
-        }
-
-        if (Array.isArray(set)) {
-            return this.insert_batch(table, set, ignore, suffix);
-        }
-
-        if (set !== null) {
-            if (Object.prototype.toString.call(set) === Object.prototype.toString.call({}) && Object.keys(set).length > 0) {
-                this.set(set);
-            }
-        }
-
-        if (typeof table !== 'string') {
-            throw new Error("insert(): Table parameter must be a string!");
-        }
-
-        table = table.trim();
-
-        if (table !== '' && !(/^[a-zA-Z0-9\$_]+(\.[a-zA-Z0-9\$_]+)?$/).test(table)) {
-            throw new Error("insert(): Invalid table name ('" + table + "') provided!");
-        }
-
-        if (table === '') {
-            if (this.from_array.length == 0) throw new Error('insert(): No tables set to insert into!');
-            table = this.from_array[0];
-        } else {
-            this._clear_array(this.from_array);
-            this.from(table);
-        }
-        return this._compile_insert(ignore, suffix);
-    }
-
-    insert_ignore(table, set, suffix) {
-        return this._insert_ignore(table, set, suffix);
-    }
-
-    _insert_ignore(table, set, suffix) {
-        return this.insert(table, set, true, suffix);
-    }
-
-    insert_batch(table,set=null,ignore,suffix) {
-        return this._insert_batch(table, set, ignore, suffix);
-    }
-
     _insert_batch(table,set=null,ignore,suffix) {
         const orig_table = table = table || '';
         ignore = (typeof ignore !== 'boolean' ? false : ignore);
@@ -1334,7 +393,7 @@ class QueryBuilder {
                     }
                 }
                 if (row.length != columns.length) {
-                    throw new Error('insert_batch(): Cannot use batch insert into ' + table + ' - fields must match on all rows (' + row.join(',') + ' vs ' + columns.join(',') + ').');
+                    throw new Error(`insert_batch(): Cannot use batch insert into ${table} - fields must match on all rows (${row.join(',')} vs ${columns.join(',')}).`);
                 }
                 map.push('(' + row.join(', ') + ')');
             })(i);
@@ -1342,59 +401,6 @@ class QueryBuilder {
 
         const verb = 'INSERT ' + (ignore === true ? 'IGNORE ' : '');
         return verb + `INTO ${this.from_array[0]} (${columns.join(', ')}) VALUES ${map.join(', ')}${suffix}`;
-    }
-
-    get(table) {
-        return this._get(table);
-    }
-
-    _get(table) {
-        if (typeof table === 'string' || Array.isArray(table)) {
-            this.from(table);
-        }
-        else {
-            if (this.from_array.length == 0) {
-                throw new Error('You have not specified any tables to select from!');
-            }
-        }
-        return this._compile_select(this);
-    }
-
-    get_where(table=null, where=null) {
-        return this._get_where(table, where);
-    }
-
-    _get_where(table=null, where=null) {
-        // Check if table is either a string or array
-        if (typeof table !== 'string' && !Array.isArray(table))
-            throw new Error('You must specify a table or array of tables in the first parameter of get_where()');
-
-        // If table is a string, make sure it's not empty
-        if (typeof table === 'string' && table.trim().length <= 0)
-            throw new Error("Invalid table string specified!");
-
-        // If table is array, make sure there are only strings in there and that they are non-empty strings
-        if (Array.isArray(table)) {
-            for (let v in table) {
-                if (typeof v !== 'string' || (typeof v === 'string' && v.trim().length <= 0)) {
-                    throw new Error("Invalid table string specified in array of tables!");
-                    break;
-                }
-            }
-        }
-
-        this.from(table);
-
-        if (where === null || typeof where !== 'object' || Object.keys(where).length === 0)
-            throw new Error('You must supply an object of field:value pairs in the second parameter of get_where()');
-
-        this.where(where);
-
-        return this._compile_select();
-    }
-
-    count(table) {
-        return this._count(table);
     }
 
     _count(table) {
@@ -1409,333 +415,7 @@ class QueryBuilder {
 
         return sql;
     }
-
-    update(table, set, where=null) {
-        return this._update(table, set, where);
-    }
-
-    _update(table, set, where=null) {
-
-        table = table || '';
-        set = set || null;
-
-        // Send to batch_update if the data param is an array
-        if (Array.isArray(set)) {
-            let index = null;
-            if (set.length > 0) {
-                if (Object.prototype.toString.call(set[0]) === Object.prototype.toString.call({})) {
-                    index = Object.keys(set[0])[0];
-                }
-            }
-            if (index) {
-                return this.update_batch(table, set, index, where);
-            } else {
-                throw new Error("update(): update_batch attempted but could not ascertain a valid index to use from the dataset provided.");
-            }
-        }
-
-        // If set is a number, boolean, a non-empty string, or regex, fail
-        if (/^(number|boolean)$/.test(typeof set) || (typeof set == 'string' && set !== '') || Object.prototype.toString.call(set) === Object.prototype.toString.call(/test/)) {
-            throw new Error("update(): Invalid data provided to update database!");
-        }
-
-        // If data object was provided, set it
-        if (set !== null) {
-            if (Object.prototype.toString.call(set) === Object.prototype.toString.call({}) && Object.keys(set).length > 0) {
-                this.set(set);
-            } else {
-                throw new Error("update(): Empty data object provided. This is not allowed.");
-            }
-        }
-
-        // Fail if, at this point, nothing has been set
-        if (this.set_array.length == 0) {
-            throw new Error("update(): You must set some field value pairs to update using the set method or via an object passed to the second parameter of the update method!");
-        }
-
-        // NOTE: If falsy table provided, table will have been converted to an empty string...
-        if (typeof table !== 'string') {
-            throw new Error("update(): Table parameter must be a string!");
-        }
-
-        table = table.trim();
-
-        // Table name must be in a legitimate format
-        if (table !== '' && !/^[a-zA-Z0-9\$_]+(\.[a-zA-Z0-9\$_]+)?$/.test(table)) {
-            throw new Error("update(): You have not set any tables to update!");
-        }
-
-        // If table not supplied, it must have been supplied already
-        if (table == '') {
-            if (this.from_array.length == 0) {
-                throw new Error('update(): No tables set to update!');
-            }
-            table = this.from_array[0];
-        } else {
-            this._clear_array(this.from_array);
-            this.from(table);
-        }
-
-        // Set where array if a where statement was provided
-        if (where !== null) {
-            this.where(where);
-        }
-
-        return this._compile_update();
-    }
-
-    update_batch(table='', set=null, index=null, where=null) {
-        return this._update_batch(table, set, index, where);
-    }
-
-    _update_batch(table='', set=null, index=null, where=null) {
-        // Make sure an index has been provided!
-        if (typeof index !== 'string' || (typeof index === 'string' && index.length === 0)) {
-            throw new Error("update_batch(): Invalid index provided to generate batch update query!");
-        }
-
-        // Check to make sure we have a dataset
-        if (!Array.isArray(set)) {
-            throw new Error("update_batch(): Array of object expected and non-array received.");
-        }
-
-        // Make sure our dataset isn't emtpy
-        if (set.length === 0) {
-            throw new Error("update_batch(): You must supply some data to batch update the table with.");
-        }
-
-        // Make sure each item in the dataset has the specified index and then add data to set_array
-        //console.dir(set);
-        for (let i in set) {
-            const clean = {};
-            const row = set[i];
-            if (Object.prototype.toString.call(row) === Object.prototype.toString.call({}) && Object.keys(row).length > 0) {
-                const keys = Object.keys(row);
-                if (keys.indexOf(index) !== -1) {
-                    for (let j in row) {
-                        clean[this._protect_identifiers(j)] = this._qb_escape(row[j]);
-                    }
-                    this.set_array.push(clean);
-                }
-            } else {
-                throw new Error("update_batch(): You have supplied an invalid object to batch update!");
-            }
-        }
-
-        // Fail if, at this point, nothing has been set
-        if (this.set_array.length == 0) {
-            throw new Error("update_batch(): You must provide some data to batch update!");
-        }
-
-        // NOTE: If falsy table provided, table will have been converted to an empty string...
-        if (typeof table !== 'string') {
-            throw new Error("update(): Table parameter must be a string!");
-        }
-
-        table = table.trim();
-
-        // Table name must be in a legitimate format
-        if (table !== '' && !/^[a-zA-Z0-9\$_]+(\.[a-zA-Z0-9\$_]+)?$/.test(table)) {
-            throw new Error("update(): You have not set any tables to update!");
-        }
-
-        // If table not supplied, it must have been supplied already
-        if (table == '') {
-            if (this.from_array.length == 0) {
-                throw new Error('No tables set to insert into!');
-            }
-            table = this.from_array[0];
-        } else {
-            this._clear_array(this.from_array);
-            this.from(table);
-        }
-
-        // Set where array if a where statement was provided
-        if (where != null) {
-            this.where(where);
-        }
-
-        // Verify there is a table in the from_array
-        if (this.from_array.length !== 1) {
-            if (this.from_array.length === 0) {
-                throw new Error("You haven't provided any tables to build batch UPDATE query with!");
-            }
-            throw new Error("You have provided too many tables to build batch UPDATE query with!");
-        }
-
-        table = this.from_array.toString();
-
-
-        // Limit to 100 rows per batch
-        const batches = [];
-        for (let i = 0, total = this.set_array.length; i < total; i += 100) {
-            const when_then = {};
-            const ids = [];
-            const where = (this.where_array.length > 0 ? this._build_where_clause() + ' AND ' : '');
-            const chunk = this.set_array.slice(i,100);
-
-            // Escape the index
-            index = this._protect_identifiers(index);
-
-            for (let j in chunk) {
-                ids.push(chunk[j][index]);
-
-                const keys = Object.keys(chunk[j]);
-                for (let k in keys) {
-                    if (keys[k] != index) {
-                        if (!when_then.hasOwnProperty(keys[k])) {
-                            when_then[keys[k]] = [];
-                        }
-                        when_then[keys[k]].push(`WHEN ${index} = ${ids[j]} THEN ${chunk[j][keys[k]]} `);
-                    }
-                }
-            }
-
-            // Build the actual SQL statement
-            let sql = `UPDATE (${table}) SET `;
-            let cases = '';
-
-            for (let l in when_then) {
-                cases += l + ' = CASE ';
-
-                for (let m in when_then[l]) {
-                    cases += when_then[l][m];
-                }
-
-                cases += `ELSE ${l} END, `;
-            }
-
-            sql += cases.substr(0, cases.length - 2);
-            sql += ' WHERE ' + where + index + ' IN (' + ids.join(',') + ')';
-
-            // Add query to batch
-            batches.push(sql);
-        }
-
-        return batches;
-    }
-
-    delete(table, where) {
-        return this._delete(table, where);
-    }
-
-    _delete(table, where) {
-        if (typeof table == 'string' && table.trim().length > 0) {
-            this._clear_array(this.from_array);
-            this.from(table);
-        }
-
-        if (Object.prototype.toString.call(where) === Object.prototype.toString.call({})) {
-            if (Object.keys(where).length == 0) {
-                throw new Error("where(): The object you provided to limit the deletion of rows is empty.");
-            }
-            else {
-                this.where(where);
-            }
-        }
-
-        return this._compile_delete();
-    }
-
-    get_compiled_select(table) {
-        if (typeof table !== 'undefined') {
-            this._track_aliases(table);
-            this.from(table);
-        }
-        else {
-            if (this.from_array.length == 0) {
-                throw new Error('You have not specified any tables to build a select statement with!');
-                return this;
-            }
-        }
-
-        return this._compile_select();
-    }
-
-    get_compiled_delete(table) {
-        if (typeof table !== 'function') {
-            this._track_aliases(table);
-            this.from(table);
-        }
-
-        return this._compile_delete();
-    }
-
-    get_compiled_update(table) {
-        if (typeof table !== 'function') {
-            this._track_aliases(table);
-            this.from(table);
-        }
-        return this._compile_update();
-    }
-
-    get_compiled_insert(table) {
-        if (typeof table !== 'function') {
-            this._track_aliases(table);
-            this.from(table);
-        }
-        return this._compile_insert();
-    }
-
-    compile_select(table) {
-        return this.get_compiled_select(table);
-    }
-
-    compile_delete(table) {
-        return this.get_compiled_delete(table);
-    }
-
-    compile_update(table) {
-        return this.get_compiled_update(table);
-    }
-
-    compile_insert(table) {
-        return this.get_compiled_insert(table);
-    }
-
-    last_query() {
-        return this.last_query_string[0] || '';
-    }
-
-    escape(val) {
-        return this._qb_escape(val);
-    }
-
-    empty_table(table) {
-        return this._empty_table(table);
-    }
-
-    _empty_table(table) {
-        if (typeof table === 'string' && table.trim().length > 0) {
-            this._clear_array(this.from_array);
-            this.from(table);
-        }
-
-        if (this.from_array.length === 0) {
-            throw new Error('empty_table(): You have not specified a table to empty!');
-            return '';
-        }
-
-        return "DELETE FROM " + this.from_array[0];
-    }
-
-    truncate(table) {
-        return this._truncate(table);
-    }
-
-    _truncate(table) {
-        if (typeof table === 'string' && table.trim().length > 0) {
-            this._clear_array(this.from_array);
-            this.from(table);
-        }
-
-        if (this.from_array.length === 0) {
-            throw new Error('truncate(): You have not specified a table to truncate!');
-            return '';
-        }
-
-        return "TRUNCATE " + this.from_array[0];
-    }
 }
+
 
 module.exports = QueryBuilder;
